Handle `error` response when doing an auth call
It turned out that sometimes the API answers with an `error` instead of `errors` and `type`. It's not really returning a message,
so I proposed to just take the error in the message, even though it looks more like a `type` than a `Message`.

It's necessary to handle it because otherwise users just get an empty ResponseError object, with not a single idea of what's going wrong.
Maybe it's just needed for the API to answer with similar response than other calls, and this diff would be useless then :)