Add snakify_response_keys configuration
This commit introduces a new feature that allows users to configure the response hash keys to be in snake_case instead of CamelCase, following the ruby convention. Adding a dependency from 'activesupport', this increases the minimum ruby version to 2.2.2

The new feature is implemented in a way that it does not break existing functionality. There are also additional tests to ensure the correctness of the new feature.